streamlit/requirements.txt
streamlit==1.41.0
snowflake-connector-python==3.12.3
plotly==5.24.1
pandas==2.2.3
